split quick_options between classes re #186
